requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
google-generativeai==0.3.2
requests==2.31.0
reportlab==4.0.7
pyppeteer==1.0.2 